requirements.txt
streamlit==1.41.1
langgraph==0.2.53
langchain==0.3.14
langchain-groq>=0.2.0
langchain-google-genai==2.0.8
google-generativeai==0.8.3
pydantic==2.10.4
python-dotenv==1.0.1
